Nombre del Proyecto: Siva Diamantes F2 Comentarios: Avance corrigen los comandos factoreDepreciacion y rangoPesos.

+++ src/main/java/mx/com/nmp/ms/sivad/referencia/infrastructure/jpa/repository/FactorDepreciacionRepositoryImpl.java
@@ -3,6 +3,9 @@ package mx.com.nmp.ms.sivad.referencia.infrastructure.jpa.repository;
 import java.math.BigDecimal;
 import java.util.List;
 import javax.inject.Inject;
+
+import mx.com.nmp.ms.sivad.referencia.dominio.exception.FactorDepreciacionNoEncontradoException;
+import org.joda.time.DateTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
@@ -22,18 +25,28 @@ public class FactorDepreciacionRepositoryImpl implements FactorDepreciacionRepos
 
 	@Override
 	public FactorDepreciacionDiamanteJPA save(FactorDepreciacionDiamanteJPA elemento) {
+        elemento.setFecha(DateTime.now());
 		return factorDepreciacionJPARepository.save(elemento);
 	}
 
-	
+
 	public FactorDepreciacionDiamanteJPA update(Long idFactor, BigDecimal factor) {
-		
+
 		FactorDepreciacionDiamanteJPA pq = this.factorDepreciacionJPARepository.findOne(idFactor);
+
+        if(ObjectUtils.isEmpty(pq)) {
+            String msg = "El catalogo FactorDepreciacion no contiene un elemento con el identificador [" + idFactor + "].";
+            throw new FactorDepreciacionNoEncontradoException(FactorDepreciacionDiamanteJPA.class, msg);
+
+        }
+
         if (ObjectUtils.isEmpty(factor)) {
             LOGGER.warn("{}.factor = null, se deja valor anterior {}");
         } else {
             pq.setFactor(factor);
         }
+
+        pq.setFecha(DateTime.now());
 
 		return factorDepreciacionJPARepository.saveAndFlush(pq);
 	}
@@ -41,9 +54,15 @@ public class FactorDepreciacionRepositoryImpl implements FactorDepreciacionRepos
 
 	@Override
 	public FactorDepreciacionDiamanteJPA delete(Long idFactor) {
-		FactorDepreciacionDiamanteJPA fd = this.factorDepreciacionJPARepository.getOne(idFactor);
+		FactorDepreciacionDiamanteJPA fd = this.factorDepreciacionJPARepository.findOne(idFactor);
+
+        if(ObjectUtils.isEmpty(fd)) {
+            String mensaje = "El catalogo FactorDepreciacion no contiene un elemento con el identificador [" + idFactor + "].";
+            throw new FactorDepreciacionNoEncontradoException(FactorDepreciacionDiamanteJPA.class, mensaje);
+        }
+
 		factorDepreciacionJPARepository.delete(fd);
-		
+
 		return fd;
 	}
 
@@ -51,5 +70,9 @@ public class FactorDepreciacionRepositoryImpl implements FactorDepreciacionRepos
 	public List<FactorDepreciacionDiamanteJPA> getAll() {
 		return factorDepreciacionJPARepository.findAll();
 	}
+
+    public FactorDepreciacionDiamanteJPA obtenerElemento(Long idFactor) {
+        return factorDepreciacionJPARepository.findOne(idFactor);
+    }
 
 }
